project/25-SS-Bobby/Integrated Application/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
pillow==11.3.0
PyMySQL==1.1.2
lxml
sqlparse==0.5.3
pycountry
